Fix channelmask for AC3/DTS decoding with libavcodec
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@4362 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/audioFilters/TaudioFilterMixer.h
@@ -16,7 +16,7 @@ CCritSec csMatrix;
 const TmixerMatrix::mixer_matrix_t *matrixPtr;
 int inmask,outmask;
 unsigned int oldnchannels;
-uint64_t oldchannelmask;
+DWORD oldchannelmask;
 int oldsf;
 TmixerSettings oldcfg;
